extensive code reorganization
