+++ inetdiag/inetdiag.go
@@ -31,38 +31,7 @@ expressed in host-byte order"
 */
 
 import (
-	"bytes"
-	"encoding/binary"
-	"errors"
-	"fmt"
-	"io"
-	"log"
-	"net"
-	"reflect"
 	"syscall"
-	"time"
-	"unsafe"
-
-	"golang.org/x/sys/unix"
-
-	"github.com/m-lab/tcp-info/tcp"
-)
-
-// TODO: Refactor this package, or at least this file. It feels like it
-// currently defines a little too much, does a little too much, and exports a
-// little too much. Peter and Greg each suspect there are, implicitly, multiple
-// packages defined in this file and/or directory.
-
-// Error types.
-var (
-	ErrParseFailed = errors.New("Unable to parse InetDiagMsg")
-	ErrNotType20   = errors.New("NetlinkMessage wrong type")
-)
-
-// Constants from linux.
-const (
-	TCPDIAG_GETSOCK     = 18 // uapi/linux/inet_diag.h
-	SOCK_DIAG_BY_FAMILY = 20 // uapi/linux/sock_diag.h
 )
 
 // inet_diag.h
@@ -93,413 +62,4 @@ const (
 var diagFamilyMap = map[uint8]string{
 	syscall.AF_INET:  "tcp",
 	syscall.AF_INET6: "tcp6",
-}
-
-// InetDiagSockID is the binary linux representation of a socket, as in linux/inet_diag.h
-// Linux code comments indicate this struct uses the network byte order!!!
-type InetDiagSockID struct {
-	IDiagSPort [2]byte
-	IDiagDPort [2]byte
-	IDiagSrc   [16]byte
-	IDiagDst   [16]byte
-	IDiagIf    [4]byte
-	// TODO - change this to [2]uint32 ?
-	IDiagCookie [8]byte
-}
-
-// Interface returns the interface number.
-func (id *InetDiagSockID) Interface() uint32 {
-	return binary.BigEndian.Uint32(id.IDiagIf[:])
-}
-
-// SrcIP returns a golang net encoding of source address.
-func (id *InetDiagSockID) SrcIP() net.IP {
-	return ip(id.IDiagSrc)
-}
-
-// DstIP returns a golang net encoding of destination address.
-func (id *InetDiagSockID) DstIP() net.IP {
-	return ip(id.IDiagDst)
-}
-
-// SPort returns the host byte ordered port.
-// In general, Netlink is supposed to use host byte order, but this seems to be an exception.
-// Perhaps Netlink is reading a tcp stack structure that holds the port in network byte order.
-func (id *InetDiagSockID) SPort() uint16 {
-	return binary.BigEndian.Uint16(id.IDiagSPort[:])
-}
-
-// DPort returns the host byte ordered port.
-// In general, Netlink is supposed to use host byte order, but this seems to be an exception.
-// Perhaps Netlink is reading a tcp stack structure that holds the port in network byte order.
-func (id *InetDiagSockID) DPort() uint16 {
-	return binary.BigEndian.Uint16(id.IDiagDPort[:])
-}
-
-// Cookie returns the SockID's 64 bit unsigned cookie.
-func (id *InetDiagSockID) Cookie() uint64 {
-	// This is a socket UUID generated within the kernel, and is therefore in host byte order.
-	return binary.LittleEndian.Uint64(id.IDiagCookie[:])
-}
-
-// TODO should use more net.IP code instead of custom code.
-func ip(bytes [16]byte) net.IP {
-	if isIpv6(bytes) {
-		return ipv6(bytes)
-	}
-	return ipv4(bytes)
-}
-
-func isIpv6(original [16]byte) bool {
-	for i := 4; i < 16; i++ {
-		if original[i] != 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func ipv4(original [16]byte) net.IP {
-	return net.IPv4(original[0], original[1], original[2], original[3]).To4()
-}
-
-func ipv6(original [16]byte) net.IP {
-	return original[:]
-}
-
-func (id *InetDiagSockID) String() string {
-	return fmt.Sprintf("%s:%d -> %s:%d", id.SrcIP().String(), id.SPort(), id.DstIP().String(), id.DPort())
-}
-
-// InetDiagReqV2 is the Netlink request struct, as in linux/inet_diag.h
-// Note that netlink messages use host byte ordering, unless NLA_F_NET_BYTEORDER flag is present.
-type InetDiagReqV2 struct {
-	SDiagFamily   uint8
-	SDiagProtocol uint8
-	IDiagExt      uint8
-	Pad           uint8
-	IDiagStates   uint32
-	ID            InetDiagSockID
-}
-
-// SizeofInetDiagReqV2 is the size of the struct.
-// TODO should we just make this explicit in the code?
-const SizeofInetDiagReqV2 = int(unsafe.Sizeof(InetDiagReqV2{})) // Should be 0x38
-
-// Serialize is provided for json serialization?
-// TODO - should use binary functions instead?
-func (req *InetDiagReqV2) Serialize() []byte {
-	return (*(*[SizeofInetDiagReqV2]byte)(unsafe.Pointer(req)))[:]
-}
-
-// Len is provided for json serialization?
-func (req *InetDiagReqV2) Len() int {
-	return SizeofInetDiagReqV2
-}
-
-// NewInetDiagReqV2 creates a new request.
-func NewInetDiagReqV2(family, protocol uint8, states uint32) *InetDiagReqV2 {
-	return &InetDiagReqV2{
-		SDiagFamily:   family,
-		SDiagProtocol: protocol,
-		IDiagStates:   states,
-	}
-}
-
-// InetDiagMsg is the linux binary representation of a InetDiag message header, as in linux/inet_diag.h
-// Note that netlink messages use host byte ordering, unless NLA_F_NET_BYTEORDER flag is present.
-type InetDiagMsg struct {
-	IDiagFamily  uint8
-	IDiagState   uint8
-	IDiagTimer   uint8
-	IDiagRetrans uint8
-	ID           InetDiagSockID
-	IDiagExpires uint32
-	IDiagRqueue  uint32
-	IDiagWqueue  uint32
-	IDiagUID     uint32
-	IDiagInode   uint32
-}
-
-func (msg *InetDiagMsg) String() string {
-	return fmt.Sprintf("%s, %s, %s", diagFamilyMap[msg.IDiagFamily], tcp.State(msg.IDiagState), msg.ID.String())
-}
-
-// RawInetDiagMsg holds the []byte representation of an InetDiagMsg
-type RawInetDiagMsg []byte
-
-// Parse returns the InetDiagMsg itself
-// Modified from original to also return attribute data array.
-func (raw RawInetDiagMsg) Parse() (*InetDiagMsg, error) {
-	align := rtaAlignOf(int(unsafe.Sizeof(InetDiagMsg{})))
-	if len(raw) < align {
-		return nil, ErrParseFailed
-	}
-	return (*InetDiagMsg)(unsafe.Pointer(&raw[0])), nil
-}
-
-func splitInetDiagMsg(data []byte) (RawInetDiagMsg, []byte) {
-	align := rtaAlignOf(int(unsafe.Sizeof(InetDiagMsg{})))
-	if len(data) < align {
-		log.Println("Wrong length", len(data), "<", align)
-		log.Println(data)
-		return nil, nil
-	}
-	return RawInetDiagMsg(data[:align]), data[align:]
-}
-
-// RawNlMsgHdr contains a byte slice version of a syscall.NlMsgHdr
-type RawNlMsgHdr []byte
-
-// Parse returns the syscall.NlMsghdr
-func (raw RawNlMsgHdr) Parse() (*syscall.NlMsghdr, error) {
-	size := int(unsafe.Sizeof(syscall.NlMsghdr{}))
-	if len(raw) != size {
-		return nil, ErrParseFailed
-	}
-	return (*syscall.NlMsghdr)(unsafe.Pointer(&raw[0])), nil
-}
-
-// Metadata contains the metadata for a particular TCP stream.
-type Metadata struct {
-	UUID      string
-	Sequence  int
-	StartTime time.Time
-}
-
-// RouteAttrValue is the type of RouteAttr.Value
-type RouteAttrValue []byte
-
-// ParsedMessage is a container for parsed InetDiag messages and attributes.
-type ParsedMessage struct {
-	// Timestamp should be truncated to 1 millisecond for best compression.
-	// Using int64 milliseconds instead reduces compressed size by 0.5 bytes/record, or about 1.5%
-	Timestamp time.Time `json:",omitempty"`
-
-	// Storing the RawIDM instead of the parsed InetDiagMsg reduces Marshalling by 2.6 usec, and
-	// typical compressed size by 3-4 bytes/record
-	RawIDM RawInetDiagMsg `json:",omitempty"` // RawInetDiagMsg within NLMsg
-	// Saving just the .Value fields reduces Marshalling by 1.9 usec.
-	Attributes []RouteAttrValue `json:",omitempty"` // RouteAttr.Value, backed by NLMsg
-	Metadata   *Metadata        `json:",omitempty"`
-}
-
-// ChangeType indicates why a new record is worthwhile saving.
-type ChangeType int
-
-// Constants to describe the degree of change between two different ParsedMessages.
-const (
-	NoMajorChange        ChangeType = iota
-	IDiagStateChange                // The IDiagState changed
-	NoTCPInfo                       // There is no TCPInfo attribute
-	NewAttribute                    // There is a new attribute
-	LostAttribute                   // There is a dropped attribute
-	AttributeLength                 // The length of an attribute changed
-	StateOrCounterChange            // One of the early fields in DIAG_INFO changed.
-	PacketCountChange               // One of the packet/byte/segment counts (or other late field) changed
-	PreviousWasNil                  // The previous message was nil
-	Other                           // Some other attribute changed
-)
-
-// Useful offsets for Compare
-const (
-	lastDataSentOffset = unsafe.Offsetof(syscall.TCPInfo{}.Last_data_sent)
-	pmtuOffset         = unsafe.Offsetof(syscall.TCPInfo{}.Pmtu)
-)
-
-// Compare compares important fields to determine whether significant updates have occurred.
-// We ignore a bunch of fields:
-//  * The TCPInfo fields matching last_* are rapidly changing, but don't have much significance.
-//    Are they elapsed time fields?
-//  * The InetDiagMsg.Expires is also rapidly changing in many connections, but also seems
-//    unimportant.
-//
-// Significant updates are reflected in the packet, segment and byte count updates, so we
-// generally want to record a snapshot when any of those change.  They are in the latter
-// part of the linux struct, following the pmtu field.
-//
-// The simplest test that seems to tell us what we care about is to look at all the fields
-// in the TCPInfo struct related to packets, bytes, and segments.  In addition to the TCPState
-// and CAState fields, these are probably adequate, but we also check for new or missing attributes
-// and any attribute difference outside of the TCPInfo (INET_DIAG_INFO) attribute.
-func (pm *ParsedMessage) Compare(previous *ParsedMessage) (ChangeType, error) {
-	if previous == nil {
-		return PreviousWasNil, nil
-	}
-	// If the TCP state has changed, that is important!
-	prevIDM, err := previous.RawIDM.Parse()
-	if err != nil {
-		return NoMajorChange, ErrParseFailed
-	}
-	pmIDM, err := pm.RawIDM.Parse()
-	if err != nil {
-		return NoMajorChange, ErrParseFailed
-	}
-	if prevIDM.IDiagState != pmIDM.IDiagState {
-		return IDiagStateChange, nil
-	}
-
-	// TODO - should we validate that ID matches?  Otherwise, we shouldn't even be comparing the rest.
-
-	a := previous.Attributes[INET_DIAG_INFO]
-	b := pm.Attributes[INET_DIAG_INFO]
-	if a == nil || b == nil {
-		return NoTCPInfo, nil
-	}
-
-	// If any of the byte/segment/package counters have changed, that is what we are most
-	// interested in.
-	if 0 != bytes.Compare(a[pmtuOffset:], b[pmtuOffset:]) {
-		return StateOrCounterChange, nil
-	}
-
-	// Check all the earlier fields, too.  Usually these won't change unless the counters above
-	// change, but this way we won't miss something subtle.
-	if 0 != bytes.Compare(a[:lastDataSentOffset], b[:lastDataSentOffset]) {
-		return StateOrCounterChange, nil
-	}
-
-	// If any attributes have been added or removed, that is likely significant.
-	for tp := range previous.Attributes {
-		switch tp {
-		case INET_DIAG_INFO:
-			// Handled explicitly above.
-		default:
-			// Detect any change in anything other than INET_DIAG_INFO
-			a := previous.Attributes[tp]
-			b := pm.Attributes[tp]
-			if a == nil && b != nil {
-				return NewAttribute, nil
-			}
-			if a != nil && b == nil {
-				return LostAttribute, nil
-			}
-			if a == nil && b == nil {
-				continue
-			}
-			if len(a) != len(b) {
-				return AttributeLength, nil
-			}
-			// All others we want to be identical
-			if 0 != bytes.Compare(a, b) {
-				return Other, nil
-			}
-		}
-	}
-
-	return NoMajorChange, nil
-}
-
-func isLocal(addr net.IP) bool {
-	return addr.IsLoopback() || addr.IsLinkLocalUnicast() || addr.IsMulticast() || addr.IsUnspecified()
-}
-
-func slice(hp *syscall.NlMsghdr) []byte {
-	hdrSlice := make([]byte, int(unsafe.Sizeof(*hp)), int(unsafe.Sizeof(*hp)))
-	hdr := (*reflect.SliceHeader)(unsafe.Pointer(&hdrSlice))
-	hdr.Data = uintptr(unsafe.Pointer(hp))
-	return hdrSlice
-}
-
-// Parse parses the NetlinkMessage into a ParsedMessage.  If skipLocal is true, it will return nil for
-// loopback, local unicast, multicast, and unspecified connections.
-// Note that Parse does not populate the Timestamp field, so caller should do so.
-func Parse(msg *syscall.NetlinkMessage, skipLocal bool) (*ParsedMessage, error) {
-	if msg.Header.Type != 20 {
-		return nil, ErrNotType20
-	}
-	raw, attrBytes := splitInetDiagMsg(msg.Data)
-	if raw == nil {
-		return nil, ErrParseFailed
-	}
-	if skipLocal {
-		idm, err := raw.Parse()
-		if err != nil {
-			return nil, err
-		}
-
-		if isLocal(idm.ID.SrcIP()) || isLocal(idm.ID.DstIP()) {
-			return nil, nil
-		}
-	}
-
-	parsedMsg := ParsedMessage{RawIDM: raw}
-	// parsedMsg.NLMsgHdr = &msg.Header
-
-	attrs, err := ParseRouteAttr(attrBytes)
-	if err != nil {
-		return nil, err
-	}
-	maxAttrType := uint16(0)
-	for _, a := range attrs {
-		t := a.Attr.Type
-		if t > maxAttrType {
-			maxAttrType = t
-		}
-	}
-	if maxAttrType > 2*INET_DIAG_MAX {
-		maxAttrType = 2 * INET_DIAG_MAX
-	}
-	parsedMsg.Attributes = make([]RouteAttrValue, maxAttrType+1, maxAttrType+1)
-	for _, a := range attrs {
-		t := a.Attr.Type
-		if t > maxAttrType {
-			log.Println("Error!! Received RouteAttr with very large Type:", t)
-			continue
-		}
-		parsedMsg.Attributes[t] = a.Value
-	}
-	return &parsedMsg, nil
-}
-
-// LoadNext is a simple utility to read the next NetlinkMessage from a source reader,
-// e.g. from a file of saved netlink messages.
-func LoadNext(rdr io.Reader) (*syscall.NetlinkMessage, error) {
-	var header syscall.NlMsghdr
-	// TODO - should we pass in LittleEndian as a parameter?
-	err := binary.Read(rdr, binary.LittleEndian, &header)
-	if err != nil {
-		// Note that this may be EOF
-		return nil, err
-	}
-	//log.Printf("%+v\n", header)
-	data := make([]byte, header.Len-uint32(binary.Size(header)))
-	err = binary.Read(rdr, binary.LittleEndian, data)
-	if err != nil {
-		return nil, err
-	}
-
-	return &syscall.NetlinkMessage{Header: header, Data: data}, nil
-}
-
-/*********************************************************************************************/
-/*             Copied from "github.com/vishvananda/netlink/nl/nl_linux.go"                   */
-/*********************************************************************************************/
-
-// ParseRouteAttr parses a byte array into a NetlinkRouteAttr struct.
-func ParseRouteAttr(b []byte) ([]syscall.NetlinkRouteAttr, error) {
-	var attrs []syscall.NetlinkRouteAttr
-	for len(b) >= unix.SizeofRtAttr {
-		a, vbuf, alen, err := netlinkRouteAttrAndValue(b)
-		if err != nil {
-			return nil, err
-		}
-		ra := syscall.NetlinkRouteAttr{Attr: syscall.RtAttr(*a), Value: vbuf[:int(a.Len)-unix.SizeofRtAttr]}
-		attrs = append(attrs, ra)
-		b = b[alen:]
-	}
-	return attrs, nil
-}
-
-// rtaAlignOf rounds the length of a netlink route attribute up to align it properly.
-func rtaAlignOf(attrlen int) int {
-	return (attrlen + unix.RTA_ALIGNTO - 1) & ^(unix.RTA_ALIGNTO - 1)
-}
-
-func netlinkRouteAttrAndValue(b []byte) (*unix.RtAttr, []byte, int, error) {
-	a := (*unix.RtAttr)(unsafe.Pointer(&b[0]))
-	if int(a.Len) < unix.SizeofRtAttr || int(a.Len) > len(b) {
-		return nil, nil, 0, unix.EINVAL
-	}
-	return a, b[unix.SizeofRtAttr:], rtaAlignOf(int(a.Len)), nil
 }
