+++ inetdiag/structs.go
@@ -1,0 +1,104 @@
+package inetdiag
+
+import (
+	"encoding/binary"
+	"fmt"
+	"net"
+
+	"github.com/m-lab/tcp-info/tcp"
+)
+
+// InetDiagSockID is the binary linux representation of a socket, as in linux/inet_diag.h
+// Linux code comments indicate this struct uses the network byte order!!!
+type InetDiagSockID struct {
+	IDiagSPort [2]byte
+	IDiagDPort [2]byte
+	IDiagSrc   [16]byte
+	IDiagDst   [16]byte
+	IDiagIf    [4]byte
+	// TODO - change this to [2]uint32 ?
+	IDiagCookie [8]byte
+}
+
+// Interface returns the interface number.
+func (id *InetDiagSockID) Interface() uint32 {
+	return binary.BigEndian.Uint32(id.IDiagIf[:])
+}
+
+// SrcIP returns a golang net encoding of source address.
+func (id *InetDiagSockID) SrcIP() net.IP {
+	return ip(id.IDiagSrc)
+}
+
+// DstIP returns a golang net encoding of destination address.
+func (id *InetDiagSockID) DstIP() net.IP {
+	return ip(id.IDiagDst)
+}
+
+// SPort returns the host byte ordered port.
+// In general, Netlink is supposed to use host byte order, but this seems to be an exception.
+// Perhaps Netlink is reading a tcp stack structure that holds the port in network byte order.
+func (id *InetDiagSockID) SPort() uint16 {
+	return binary.BigEndian.Uint16(id.IDiagSPort[:])
+}
+
+// DPort returns the host byte ordered port.
+// In general, Netlink is supposed to use host byte order, but this seems to be an exception.
+// Perhaps Netlink is reading a tcp stack structure that holds the port in network byte order.
+func (id *InetDiagSockID) DPort() uint16 {
+	return binary.BigEndian.Uint16(id.IDiagDPort[:])
+}
+
+// Cookie returns the SockID's 64 bit unsigned cookie.
+func (id *InetDiagSockID) Cookie() uint64 {
+	// This is a socket UUID generated within the kernel, and is therefore in host byte order.
+	return binary.LittleEndian.Uint64(id.IDiagCookie[:])
+}
+
+// TODO should use more net.IP code instead of custom code.
+func ip(bytes [16]byte) net.IP {
+	if isIpv6(bytes) {
+		return ipv6(bytes)
+	}
+	return ipv4(bytes)
+}
+
+func isIpv6(original [16]byte) bool {
+	for i := 4; i < 16; i++ {
+		if original[i] != 0 {
+			return true
+		}
+	}
+	return false
+}
+
+func ipv4(original [16]byte) net.IP {
+	return net.IPv4(original[0], original[1], original[2], original[3]).To4()
+}
+
+func ipv6(original [16]byte) net.IP {
+	return original[:]
+}
+
+func (id *InetDiagSockID) String() string {
+	return fmt.Sprintf("%s:%d -> %s:%d", id.SrcIP().String(), id.SPort(), id.DstIP().String(), id.DPort())
+}
+
+// InetDiagMsg is the linux binary representation of a InetDiag message header, as in linux/inet_diag.h
+// Note that netlink messages use host byte ordering, unless NLA_F_NET_BYTEORDER flag is present.
+type InetDiagMsg struct {
+	IDiagFamily  uint8
+	IDiagState   uint8
+	IDiagTimer   uint8
+	IDiagRetrans uint8
+	ID           InetDiagSockID
+	IDiagExpires uint32
+	IDiagRqueue  uint32
+	IDiagWqueue  uint32
+	IDiagUID     uint32
+	IDiagInode   uint32
+}
+
+func (msg *InetDiagMsg) String() string {
+	return fmt.Sprintf("%s, %s, %s", diagFamilyMap[msg.IDiagFamily], tcp.State(msg.IDiagState), msg.ID.String())
+}
